change to JPA: remove expense_category view
